extract restore constants
